Move parseNumberForOptions from utils to cli module
- Relocated the parseNumberForOptions function to lib/app/cli.ts
- Updated corresponding test cases
- Left other utility functions in place
- Updated imports in affected files

This completes the first phase of reorganizing utility functions
for better modularity.
